Update models with string link with other models

+++ bodzify_api/model/track/LibraryTrack.py
@@ -6,10 +6,6 @@ from django.contrib.auth.models import User
 from django.core.validators import FileExtensionValidator
 from django.core.validators import MaxValueValidator, MinValueValidator
 from upload_validator import FileTypeValidator
-from bodzify_api.model.criteria.Criteria import Criteria
-from bodzify_api.model.playlist.Playlist import Playlist
-from bodzify_api.model.Artist import Artist
-from bodzify_api.model.Album import Album
 from bodzify_api.validator.LibraryTrackSizeValidator import trackSize
 import bodzify_api.settings as settings
 
@@ -28,37 +24,45 @@ class LibraryTrack(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     file = models.FileField(
             upload_to=userDirectoryPath, 
-            help_text="Only audio formats accepted", 
+            help_text="Only audio formats accepted.", 
             validators=[
                     FileExtensionValidator(['flac', 'wav', 'mp3']), 
                     FileTypeValidator(allowed_types=[ 'audio/*']),
-                    trackSize])
+                    trackSize],
+            null=True)
     title = models.CharField(max_length=100, default=None)
-    artist = models.ForeignKey(Artist, on_delete=models.CASCADE, default=None, null=True)
-    album = models.ForeignKey(Album, on_delete=models.CASCADE, default=None, null=True)
-    genre = models.ForeignKey(Criteria, on_delete=models.DO_NOTHING)
+    artist = models.ForeignKey(
+        'bodzify_api.Artist', on_delete=models.CASCADE, default=None, null=True)
+    album = models.ForeignKey('bodzify_api.Album', on_delete=models.CASCADE, default=None, null=True)
+    genre = models.ForeignKey('bodzify_api.Criteria', on_delete=models.DO_NOTHING)
     duration = models.FloatField(default=None)
     rating = models.IntegerField(
             default=0, validators=[MinValueValidator(0), MaxValueValidator(255)])
-    playlists = models.ManyToManyField(Playlist)
+    playlists = models.ManyToManyField('bodzify_api.Playlist')
     language = models.CharField(max_length=100, default=None, null=True)
     addedOn = models.DateTimeField(auto_now_add=True, editable=False)
 
 
     @property
     def filename(self) -> str:
-        return os.path.basename(self.file.path)
+        if self.fileExists:
+            return os.path.basename(self.file.path)
+        return None
 
 
     @property
     def fileExtension(self) -> str:
-        filename, fileExtension = os.path.splitext(self.file.name)
-        return fileExtension
+        if self.fileExists:
+            filename, fileExtension = os.path.splitext(self.file.name)
+            return fileExtension
+        return None
 
 
     @property
     def fileExists(self) -> bool:
-        return os.path.isfile(self.file.path)
+        if self.file:
+            return os.path.isfile(self.file.path)
+        return False
 
 
     @property
@@ -73,7 +77,7 @@ class LibraryTrack(models.Model):
     def delete(self):
         if self.fileExists:
             self.file.delete()
-        if LibraryTrack.objects.filter(user=self.user, album=self.album).count() == 1:
-            self.album.delete()
-        else:
-            super(LibraryTrack, self).delete()
+        if self.album_id is not None:
+            if LibraryTrack.objects.filter(user=self.user, album=self.album).count() == 1:
+                self.album.delete()
+        super(LibraryTrack, self).delete()
